Feat: adicionado metot de profile
