+++ src/main/java/org/example/apifluxar/dto/employee/EmployeeResponseDTO.java
@@ -4,32 +4,21 @@ import org.example.apifluxar.dto.sector.SectorResponseDTO;
 import org.example.apifluxar.dto.unit.UnitResponseDTO;
 
 public class EmployeeResponseDTO {
-    private Long id;
     private String firstName;
     private String lastName;
     private String email;
-    private Character role;
     private String profilePhoto;
     private SectorResponseDTO sector;
     private UnitResponseDTO unit;
     private Double maxCapacity;
 
-    public EmployeeResponseDTO(Long id, String firstName, String lastName, String email, Character role, String profilePhoto) {
-        this.id = id;
+    public EmployeeResponseDTO(String firstName, String lastName, String email, String profilePhoto) {
         this.firstName = firstName;
         this.lastName = lastName;
         this.email = email;
-        this.role = role;
         this.profilePhoto = profilePhoto;
     }
 
-    public long getId() {
-        return id;
-    }
-
-    public void setId(long id) {
-        this.id = id;
-    }
 
     public String getFirstName() {
         return firstName;
@@ -67,18 +56,6 @@ public class EmployeeResponseDTO {
     }
     public void setSector(SectorResponseDTO sector) {
         this.sector = sector;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
-    }
-
-    public Character getRole() {
-        return role;
-    }
-
-    public void setRole(Character role) {
-        this.role = role;
     }
 
     public UnitResponseDTO getUnit() {
